Don't throw errors when tinyint(1) columns on mysql have non-boolean values (#97)
As mentioned on that ticket, https://dev.mysql.com/doc/relnotes/mysql/8.0/en/news-8-0-19.html says display widths are deprecated except for the special case that they assume that tinyint(1) is used for boolean columns:

	Display width specification for integer data types was deprecated in MySQL 8.0.17, and now statements that include data type definitions in their output no longer show the display width for integer types, with these exceptions:

	* The type is TINYINT(1). MySQL Connectors make the assumption that TINYINT(1) columns originated as BOOLEAN columns; this exception enables them to continue to make that assumption.

But as per the ticket, some users in legacy apps do have tinyint(1) with non-boolean values. Try not to error out on the mysql -> mysql case, at least; we can't do anything for the mysql -> postgresql case at this point, though, so that will still fail at the receiving end (testing shows this produces an error rather than incorrect data, so I can live with that).

You'd think we could add a special type or option for this, but it's quite hard to do because the column type conversion code only has access to the resultset and not any other information, so the typecast cases are effectively hardcoded.